Исправление замечаний

+++ src/main/java/ru/yandex/practicum/filmorate/storage/event/EventStorage.java
@@ -5,8 +5,6 @@ import ru.yandex.practicum.filmorate.model.Event;
 import java.util.List;
 
 public interface EventStorage {
-
     void add(Event event);
-
-    List<Event> getByUserId(Integer userId);
+    List<Event> getByUserId(Long userId);
 }
